apps/ml-backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pillow==10.1.0
numpy==1.24.3
opencv-python==4.8.1.78
torch==2.1.0
torchvision==0.16.0
transformers==4.35.0
scikit-learn==1.3.0
python-jose==3.3.0
passlib==1.7.4
python-dotenv==1.0.0
pytest==7.4.3
black==23.11.0
isort==5.12.0
flake8==6.1.0
